requirements.txt
pandas
folium
pyspark
sqlalchemy
geopandas
psycopg2-binary
streamlit
streamlit-folium